Move the traits implementations to specific modules.
