+++ src/tensor/mod.rs
@@ -1,287 +1,17 @@
 //! A Torch tensor.
-use crate::{Device, Kind, Scalar, TchError};
-use half::f16;
-use std::convert::{TryFrom, TryInto};
-use std::ops::{Add, AddAssign, Div, DivAssign, Mul, MulAssign, Neg, Sub, SubAssign};
+use crate::{Device, Kind, TchError};
 use torch_sys::*;
 
+mod convert;
 pub mod index;
 mod iter;
 mod npy;
+mod ops;
 
 pub use super::wrappers::tensor::{
     autocast, no_grad, no_grad_guard, with_grad, NoGradGuard, Reduction, Tensor,
 };
 pub use index::{IndexOp, NewAxis, TensorIndexer};
-
-macro_rules! impl_op {
-    ($trait:ident, $func:ident, $op:ident) => {
-        impl $trait<Tensor> for Tensor {
-            type Output = Tensor;
-
-            fn $func(self, rhs: Tensor) -> Self::Output {
-                self.$op(&rhs)
-            }
-        }
-
-        impl $trait<&Tensor> for Tensor {
-            type Output = Tensor;
-
-            fn $func(self, rhs: &Tensor) -> Self::Output {
-                self.$op(rhs)
-            }
-        }
-
-        impl<'a> $trait<&Tensor> for &'a Tensor {
-            type Output = Tensor;
-
-            fn $func(self, rhs: &Tensor) -> Self::Output {
-                self.$op(rhs)
-            }
-        }
-
-        impl $trait<Tensor> for &Tensor {
-            type Output = Tensor;
-
-            fn $func(self, rhs: Tensor) -> Self::Output {
-                self.$op(&rhs)
-            }
-        }
-    };
-}
-
-impl<S> Add<S> for &Tensor
-where
-    S: Into<Scalar>,
-{
-    type Output = Tensor;
-
-    fn add(self, rhs: S) -> Self::Output {
-        self.g_add1(rhs)
-    }
-}
-
-impl<S> Add<S> for Tensor
-where
-    S: Into<Scalar>,
-{
-    type Output = Tensor;
-
-    fn add(self, rhs: S) -> Self::Output {
-        (&self).add(rhs)
-    }
-}
-
-impl<S> Sub<S> for &Tensor
-where
-    S: Into<Scalar>,
-{
-    type Output = Tensor;
-
-    fn sub(self, rhs: S) -> Self::Output {
-        self.g_sub1(rhs)
-    }
-}
-
-impl<S> Sub<S> for Tensor
-where
-    S: Into<Scalar>,
-{
-    type Output = Tensor;
-
-    fn sub(self, rhs: S) -> Self::Output {
-        (&self).sub(rhs)
-    }
-}
-
-impl<S> Mul<S> for &Tensor
-where
-    S: Into<Scalar>,
-{
-    type Output = Tensor;
-
-    fn mul(self, rhs: S) -> Self::Output {
-        self.g_mul1(rhs)
-    }
-}
-
-impl<S> Mul<S> for Tensor
-where
-    S: Into<Scalar>,
-{
-    type Output = Tensor;
-
-    fn mul(self, rhs: S) -> Self::Output {
-        (&self).mul(rhs)
-    }
-}
-
-impl<S> Div<S> for &Tensor
-where
-    S: Into<Scalar>,
-{
-    type Output = Tensor;
-
-    fn div(self, rhs: S) -> Self::Output {
-        self.g_div1(rhs)
-    }
-}
-
-impl<S> Div<S> for Tensor
-where
-    S: Into<Scalar>,
-{
-    type Output = Tensor;
-
-    fn div(self, rhs: S) -> Self::Output {
-        (&self).div(rhs)
-    }
-}
-
-macro_rules! impl_op_basic {
-    /* rev such that rev(op(b, a)) = op(a, b) */
-    ($trait:ident, $func:ident, $op:ident, $rev:ident) => {
-        impl $trait<Tensor> for i32 {
-            type Output = Tensor;
-
-            fn $func(self, rhs: Tensor) -> Self::Output {
-                self.$func(&rhs)
-            }
-        }
-
-        impl $trait<Tensor> for i64 {
-            type Output = Tensor;
-
-            fn $func(self, rhs: Tensor) -> Self::Output {
-                self.$func(&rhs)
-            }
-        }
-
-        impl $trait<Tensor> for f32 {
-            type Output = Tensor;
-
-            fn $func(self, rhs: Tensor) -> Self::Output {
-                self.$func(&rhs)
-            }
-        }
-
-        impl $trait<Tensor> for f64 {
-            type Output = Tensor;
-
-            fn $func(self, rhs: Tensor) -> Self::Output {
-                self.$func(&rhs)
-            }
-        }
-
-        impl $trait<&Tensor> for i32 {
-            type Output = Tensor;
-
-            fn $func(self, rhs: &Tensor) -> Self::Output {
-                $rev(rhs.$op(self as i64))
-            }
-        }
-
-        impl $trait<&Tensor> for i64 {
-            type Output = Tensor;
-
-            fn $func(self, rhs: &Tensor) -> Self::Output {
-                $rev(rhs.$op(self))
-            }
-        }
-
-        impl $trait<&Tensor> for f32 {
-            type Output = Tensor;
-
-            fn $func(self, rhs: &Tensor) -> Self::Output {
-                $rev(rhs.$op(self as f64))
-            }
-        }
-
-        impl $trait<&Tensor> for f64 {
-            type Output = Tensor;
-
-            fn $func(self, rhs: &Tensor) -> Self::Output {
-                $rev(rhs.$op(self))
-            }
-        }
-    };
-}
-
-macro_rules! impl_op_assign {
-    ($trait:ident, $func:ident, $op:ident) => {
-        impl $trait<Tensor> for Tensor {
-            fn $func(&mut self, rhs: Tensor) {
-                let _ = self.$op(&rhs);
-            }
-        }
-
-        impl $trait<&Tensor> for Tensor {
-            fn $func(&mut self, rhs: &Tensor) {
-                let _ = self.$op(rhs);
-            }
-        }
-    };
-}
-
-macro_rules! impl_op_assign_basic {
-    ($trait:ident, $func:ident, $op:ident) => {
-        impl $trait<i32> for Tensor {
-            fn $func(&mut self, rhs: i32) {
-                let _ = self.$op(rhs as i64);
-            }
-        }
-
-        impl $trait<i64> for Tensor {
-            fn $func(&mut self, rhs: i64) {
-                let _ = self.$op(rhs);
-            }
-        }
-
-        impl $trait<f32> for Tensor {
-            fn $func(&mut self, rhs: f32) {
-                let _ = self.$op(rhs as f64);
-            }
-        }
-
-        impl $trait<f64> for Tensor {
-            fn $func(&mut self, rhs: f64) {
-                let _ = self.$op(rhs);
-            }
-        }
-    };
-}
-
-fn id<T>(v: T) -> T {
-    v
-}
-
-fn neg(t: Tensor) -> Tensor {
-    t.neg()
-}
-
-fn inv(t: Tensor) -> Tensor {
-    t.pow(-1)
-}
-
-impl_op!(Add, add, g_add);
-impl_op_basic!(Add, add, g_add1, id);
-impl_op_assign!(AddAssign, add_assign, g_add_);
-impl_op_assign_basic!(AddAssign, add_assign, g_add_1);
-
-impl_op!(Mul, mul, g_mul);
-impl_op_basic!(Mul, mul, g_mul1, id);
-impl_op_assign!(MulAssign, mul_assign, g_mul_);
-impl_op_assign_basic!(MulAssign, mul_assign, g_mul_1);
-
-impl_op!(Div, div, g_div);
-impl_op_basic!(Div, div, g_div1, inv);
-impl_op_assign!(DivAssign, div_assign, g_div_);
-impl_op_assign_basic!(DivAssign, div_assign, g_div_1);
-
-impl_op!(Sub, sub, g_sub);
-impl_op_basic!(Sub, sub, g_sub1, neg);
-impl_op_assign!(SubAssign, sub_assign, g_sub_);
-impl_op_assign_basic!(SubAssign, sub_assign, g_sub_1);
 
 pub trait Shape {
     fn to_shape(&self) -> Box<[i64]>;
@@ -403,22 +133,6 @@ impl Tensor {
     }
 }
 
-impl Neg for Tensor {
-    type Output = Tensor;
-
-    fn neg(self) -> Tensor {
-        self.f_neg().unwrap()
-    }
-}
-
-impl Neg for &Tensor {
-    type Output = Tensor;
-
-    fn neg(self) -> Tensor {
-        self.f_neg().unwrap()
-    }
-}
-
 impl<T: crate::kind::Element> From<&[T]> for Tensor {
     fn from(v: &[T]) -> Tensor {
         Tensor::of_slice(v)
@@ -482,80 +196,6 @@ impl Tensor {
         self.g_nll_loss::<Tensor>(targets, None, Reduction::Mean, -100)
     }
 }
-
-macro_rules! from_tensor {
-    ($typ:ident, $zero:expr, $kind:ident) => {
-        impl From<&Tensor> for Vec<$typ> {
-            fn from(tensor: &Tensor) -> Vec<$typ> {
-                let numel = tensor.numel();
-                let mut vec = vec![$zero; numel as usize];
-                tensor.to_kind(Kind::$kind).copy_data(&mut vec, numel);
-                vec
-            }
-        }
-
-        impl From<&Tensor> for Vec<Vec<$typ>> {
-            fn from(tensor: &Tensor) -> Vec<Vec<$typ>> {
-                let first_dim = tensor.size()[0];
-                (0..first_dim)
-                    .map(|i| Vec::<$typ>::from(tensor.get(i)))
-                    .collect()
-            }
-        }
-
-        impl From<&Tensor> for Vec<Vec<Vec<$typ>>> {
-            fn from(tensor: &Tensor) -> Vec<Vec<Vec<$typ>>> {
-                let first_dim = tensor.size()[0];
-                (0..first_dim)
-                    .map(|i| Vec::<Vec<$typ>>::from(tensor.get(i)))
-                    .collect()
-            }
-        }
-
-        impl From<&Tensor> for $typ {
-            fn from(tensor: &Tensor) -> $typ {
-                let numel = tensor.numel();
-                if numel != 1 {
-                    panic!("expected exactly one element, got {}", numel)
-                }
-                Vec::from(tensor)[0]
-            }
-        }
-
-        impl From<Tensor> for Vec<$typ> {
-            fn from(tensor: Tensor) -> Vec<$typ> {
-                Vec::<$typ>::from(&tensor)
-            }
-        }
-
-        impl From<Tensor> for Vec<Vec<$typ>> {
-            fn from(tensor: Tensor) -> Vec<Vec<$typ>> {
-                Vec::<Vec<$typ>>::from(&tensor)
-            }
-        }
-
-        impl From<Tensor> for Vec<Vec<Vec<$typ>>> {
-            fn from(tensor: Tensor) -> Vec<Vec<Vec<$typ>>> {
-                Vec::<Vec<Vec<$typ>>>::from(&tensor)
-            }
-        }
-
-        impl From<Tensor> for $typ {
-            fn from(tensor: Tensor) -> $typ {
-                $typ::from(&tensor)
-            }
-        }
-    };
-}
-
-from_tensor!(f64, 0f64, Double);
-from_tensor!(f32, 0f32, Float);
-from_tensor!(f16, f16::from_f64(0.0), Half);
-from_tensor!(i64, 0i64, Int64);
-from_tensor!(i32, 0i32, Int);
-from_tensor!(i8, 0i8, Int8);
-from_tensor!(u8, 0u8, Uint8);
-from_tensor!(bool, false, Bool);
 
 impl Tensor {
     /// Computes the cross-entropy loss based on some logits and targets.
@@ -658,98 +298,6 @@ impl Tensor {
         self.g_to_mkldnn(self.kind())
     }
 }
-
-impl std::iter::Sum for Tensor {
-    fn sum<I: Iterator<Item = Tensor>>(mut iter: I) -> Tensor {
-        match iter.next() {
-            None => Tensor::from(0.),
-            Some(t) => iter.fold(t, |acc, x| x + acc),
-        }
-    }
-}
-
-impl<'a> std::iter::Sum<&'a Tensor> for Tensor {
-    fn sum<I: Iterator<Item = &'a Tensor>>(mut iter: I) -> Tensor {
-        match iter.next() {
-            None => Tensor::from(0.),
-            Some(t) => iter.fold(t.shallow_clone(), |acc, x| x + acc),
-        }
-    }
-}
-
-impl PartialEq for Tensor {
-    fn eq(&self, other: &Tensor) -> bool {
-        if self.size() != other.size() {
-            return false;
-        }
-        match self.f_eq1(&other) {
-            Err(_) => false,
-            Ok(v) => match v.f_all() {
-                Err(_) => false,
-                Ok(v) => i64::from(v) > 0,
-            },
-        }
-    }
-}
-
-macro_rules! try_into_impl {
-    ($type:ident) => {
-        impl TryInto<ndarray::ArrayD<$type>> for &Tensor {
-            type Error = ndarray::ShapeError;
-
-            fn try_into(self) -> Result<ndarray::ArrayD<$type>, Self::Error> {
-                let v: Vec<$type> = self.into();
-                let shape: Vec<usize> = self.size().iter().map(|s| *s as usize).collect();
-                ndarray::ArrayD::from_shape_vec(ndarray::IxDyn(&shape), v)
-            }
-        }
-    };
-}
-
-try_into_impl!(f16);
-try_into_impl!(f32);
-try_into_impl!(i32);
-try_into_impl!(f64);
-try_into_impl!(i64);
-try_into_impl!(bool);
-
-macro_rules! try_from_impl {
-    ($type:ident) => {
-        impl<D> TryFrom<ndarray::Array<$type, D>> for Tensor
-        where
-            D: ndarray::Dimension,
-        {
-            type Error = TchError;
-
-            fn try_from(value: ndarray::Array<$type, D>) -> Result<Self, Self::Error> {
-                // TODO: Replace this with `?` once `std::option::NoneError` has been stabilized.
-                let slice = match value.as_slice() {
-                    None => return Err(TchError::Convert("cannot convert to slice".to_string())),
-                    Some(v) => v,
-                };
-                let tn = Self::f_of_slice(slice)?;
-                let shape: Vec<i64> = value.shape().iter().map(|s| *s as i64).collect();
-                tn.f_reshape(&shape)
-            }
-        }
-
-        impl TryFrom<Vec<$type>> for Tensor {
-            type Error = TchError;
-
-            fn try_from(value: Vec<$type>) -> Result<Self, Self::Error> {
-                let tn = Self::f_of_slice(value.as_slice())?;
-                Ok(tn)
-            }
-        }
-    };
-}
-
-try_from_impl!(f16);
-try_from_impl!(f32);
-try_from_impl!(i32);
-try_from_impl!(f64);
-try_from_impl!(i64);
-try_from_impl!(bool);
 
 #[used]
 static INIT_ARRAY: [unsafe extern "C" fn(); 1] = [dummy_cuda_dependency];
